feat(rhino): minor changes
